Added in helper methods to help make the table and also added in helper method for the drug list.
