+++ src/main/java/helper_methods/DrugListGenerator.java
@@ -1,0 +1,288 @@
+package helper_methods;
+
+public class DrugListGenerator {
+    public String[] GenerateDrugList() {
+
+        return new String[]{"Atorvastatin",
+                "Levothyroxine",
+                "Metformin",
+                "Lisinopril",
+                "Amlodipine",
+                "Metoprolol",
+                "Albuterol",
+                "Omeprazole",
+                "Losartan",
+                "Gabapentin",
+                "Hydrochlorothiazide",
+                "Sertraline",
+                "Simvastatin",
+                "Montelukast",
+                "Escitalopram",
+                "Acetaminophen",
+                "Rosuvastatin",
+                "Bupropion",
+                "Furosemide",
+                "Pantoprazole",
+                "Trazodone",
+                "Dextroamphetamine",
+                "Fluticasone",
+                "Tamsulosin",
+                "Fluoxetine",
+                "Carvedilol",
+                "Duloxetine",
+                "Meloxicam",
+                "Clopidogrel",
+                "Prednisone",
+                "Citalopram",
+                "Insulin Glargine",
+                "Potassium Chloride",
+                "Pravastatin",
+                "Tramadol",
+                "Aspirin",
+                "Alprazolam",
+                "Ibuprofen",
+                "Cyclobenzaprine",
+                "Amoxicillin",
+                "Methylphenidate",
+                "Allopurinol",
+                "Venlafaxine",
+                "Clonazepam",
+                "Ergocalciferol",
+                "Zolpidem",
+                "Apixaban",
+                "Glipizide",
+                "Spironolactone",
+                "Cetirizine",
+                "Atenolol",
+                "Oxycodone",
+                "Buspirone",
+                "Topiramate",
+                "Warfarin",
+                "Estradiol",
+                "Cholecalciferol",
+                "Lamotrigine",
+                "Quetiapine",
+                "Lorazepam",
+                "Famotidine",
+                "Folic Acid",
+                "Azithromycin",
+                "Hydroxyzine",
+                "Insulin Lispro",
+                "Diclofenac",
+                "Loratadine",
+                "Sitagliptin",
+                "Clonidine",
+                "Diltiazem",
+                "Latanoprost",
+                "Pregabalin",
+                "Doxycycline",
+                "Insulin Aspart",
+                "Amitriptyline",
+                "Paroxetine",
+                "Ondansetron",
+                "Tizanidine",
+                "Lisdexamfetamine",
+                "Rivaroxaban",
+                "Glimepiride",
+                "Propranolol",
+                "Aripiprazole",
+                "Finasteride",
+                "Naproxen",
+                "Levetiracetam",
+                "Alendronate",
+                "Fenofibrate",
+                "Dulaglutide",
+                "Oxybutynin",
+                "Celecoxib",
+                "Lovastatin",
+                "Ezetimibe",
+                "Cephalexin",
+                "Empagliflozin",
+                "Hydralazine",
+                "Mirtazapine",
+                "Cyanocobalamin",
+                "Triamcinolone",
+                "Amoxicillin; Clavulanate",
+                "Baclofen",
+                "Valproate",
+                "Tiotropium",
+                "Sumatriptan",
+                "Donepezil",
+                "Methotrexate",
+                "Isosorbide",
+                "Ferrous Sulfate",
+                "Thyroid",
+                "Acetaminophen",
+                "Valacyclovir",
+                "Esomeprazole",
+                "Valsartan",
+                "Insulin Detemir",
+                "Clindamycin",
+                "Hydroxychloroquine",
+                "Methocarbamol",
+                "Diazepam",
+                "Semaglutide",
+                "Dexmethylphenidate",
+                "Ciprofloxacin",
+                "Chlorthalidone",
+                "Rizatriptan",
+                "Nifedipine",
+                "Insulin Degludec",
+                "Norethindrone",
+                "Risperidone",
+                "Olmesartan",
+                "Morphine",
+                "Benazepril",
+                "Meclizine",
+                "Timolol",
+                "Oxcarbazepine",
+                "Liraglutide",
+                "Dicyclomine",
+                "Irbesartan",
+                "Hydrocortisone",
+                "Verapamil",
+                "Memantine",
+                "Prednisolone",
+                "Nortriptyline",
+                "Ropinirole",
+                "Benzonatate",
+                "Progesterone",
+                "Mirabegron",
+                "Methylprednisolone",
+                "Acyclovir",
+                "Docusate",
+                "Olanzapine",
+                "Nitroglycerin",
+                "Bimatoprost",
+                "Nitrofurantoin",
+                "Pioglitazone",
+                "Ketoconazole",
+                "Clobetasol",
+                "Testosterone",
+                "Azelastine",
+                "Fluconazole",
+                "Brimonidine",
+                "Desvenlafaxine",
+                "Ranitidine",
+                "Oseltamivir",
+                "Levocetirizine",
+                "Anastrozole",
+                "Phentermine",
+                "Sucralfate",
+                "Sildenafil",
+                "Mesalamine",
+                "Carbamazepine",
+                "Buprenorphine",
+                "Flecainide",
+                "Gemfibrozil",
+                "Prazosin",
+                "Lansoprazole",
+                "Diphenhydramine",
+                "Pramipexole",
+                "Ramipril",
+                "Lithium",
+                "Amiodarone",
+                "Omega-3-acid Ethyl Esters",
+                "Glyburide",
+                "Acetaminophen",
+                "Magnesium Salts",
+                "Mupirocin",
+                "Calcium",
+                "Adalimumab",
+                "Methimazole",
+                "Budesonide",
+                "Promethazine",
+                "Doxazosin",
+                "Labetalol",
+                "Terazosin",
+                "Cyclosporine",
+                "Torsemide",
+                "Medroxyprogesterone",
+                "Calcium",
+                "Dorzolamide",
+                "Dapagliflozin",
+                "Liothyronine",
+                "Sacubitril",
+                "Beclomethasone",
+                "Insulin Isophane",
+                "Metronidazole",
+                "Temazepam",
+                "Fluticasone",
+                "Erythromycin",
+                "Polyethylene",
+                "Nystatin",
+                "Cefdinir",
+                "Benztropine",
+                "Tretinoin",
+                "Mometasone",
+                "Eszopiclone",
+                "Betamethasone",
+                "Erenumab",
+                "Hydrochlorothiazide",
+                "Minocycline",
+                "Digoxin",
+                "Empagliflozin",
+                "Nebivolol",
+                "Levofloxacin",
+                "Colchicine",
+                "Ofloxacin",
+                "Vortioxetine",
+                "Linaclotide",
+                "Umeclidinium",
+                "Insulin Human",
+                "Ticagrelor",
+                "Telmisartan",
+                "Ketorolac",
+                "Hydromorphone",
+                "Epinephrine",
+                "Doxepin",
+                "Quinapril",
+                "Umeclidinium",
+                "Fexofenadine",
+                "Brimonidine;",
+                "Letrozole",
+                "Ranolazine",
+                "Lurasidone",
+                "Phenytoin",
+                "Tadalafil",
+                "Pancrelipase",
+                "Dexlansoprazole",
+                "Isotretinoin",
+                "Sodium",
+                "Solifenacin",
+                "Bisoprolol",
+                "Olopatadine",
+                "Primidone",
+                "Bumetanide",
+                "Tolterodine",
+                "Dexamethasone",
+                "Chlorhexidine",
+                "Sodium Salts",
+                "Varenicline",
+                "Zonisamide",
+                "Calcitriol",
+                "Emtricitabine;",
+                "Terbinafine",
+                "Fluocinonide",
+                "Hydrochlorothiazide",
+                "Ziprasidone",
+                "Estrogens",
+                "Sulfasalazine",
+                "Icosapent",
+                "Dexamethasone",
+                "Atomoxetine",
+                "Formoterol",
+                "Ketotifen",
+                "Bisoprolol",
+                "Sennosides",
+                "Raloxifene",
+                "Linagliptin",
+                "Canagliflozin",
+                "Alogliptin",
+                "Sotalol",
+                "Potassium",
+                "Melatonin",
+                "Isosorbide",
+                "Guanfacine"};
+    }
+}
