+++ src/main/java/main_frame/MainFrame.java
@@ -1,7 +1,7 @@
 package main_frame;
 
 import drug_fulfill.drug_fulfill_depot_inventories.DrugFulfillHelper;
-import drug_fulfill.drug_fulfill_depot_inventories.DrugFulfillScreen;
+import drug_fulfill.drug_fulfill_frameworks_drivers.DrugFulfillScreen;
 import drug_search.frameworks_drivers.DrugSearchHelper;
 import drug_search.frameworks_drivers.DrugSearchScreen;
 import site_drug_request.drug_request_frameworks_drivers.DrugRequestHelper;
