+++ src/main/java/site_drug_request/drug_request_frameworks_drivers/DrugRequestScreen.java
@@ -1,5 +1,6 @@
 package site_drug_request.drug_request_frameworks_drivers;
 
+import helper_methods.DrugListGenerator;
 import site_drug_request.drug_request_interface_adapters.DrugRequestController;
 import site_drug_request.drug_request_interface_adapters.DrugRequestPresenter;
 import site_drug_request.drug_request_interface_adapters.DrugRequestPresenterOutputBoundary;
@@ -25,9 +26,15 @@ public class DrugRequestScreen extends JPanel implements ActionListener, DrugReq
      * number of drugBottles requested by the site
      */
     JTextField drugBottle = new JTextField(15);
-//    JComboBox drugName = new JComboBox();
+    /**
+     * The response test panel after you submit a drug request
+     */
 
     JTextField requestResponse = new JTextField(40);
+    /**
+     * A drop-down menu from which you can see which drugs to select and order.
+     */
+    JComboBox<String> selectDrug;
 
     /**
      * the controller
@@ -41,8 +48,15 @@ public class DrugRequestScreen extends JPanel implements ActionListener, DrugReq
      * The Account ID of the user
      */
     String accountID;
+    /**
+     * The Drug List available in the program
+     */
+    String[] drugs;
 
-
+    /**
+     * The Top panel where you can see the drug List and submit a request
+     */
+    JPanel drugRequestPanel;
 
 
     //    public DrugRequestScreen(){
@@ -63,7 +77,6 @@ public class DrugRequestScreen extends JPanel implements ActionListener, DrugReq
         Made the requestResponse panel read only, so it can't be edited by the user
         instead it will be used to display the response to the user.
          */
-        requestResponse.setEditable(false);
 
         /*
           Packages multiple components as a single component using panel Generator.
@@ -71,7 +84,10 @@ public class DrugRequestScreen extends JPanel implements ActionListener, DrugReq
         PanelGenerator drugNamePanel = new PanelGenerator(new JLabel("Enter Drug Name"), drugName);
         PanelGenerator drugBottlePanel = new PanelGenerator(new JLabel("Enter Number of Drug Bottles"), drugBottle);
         PanelGenerator requestResponsePanel = new PanelGenerator(new JLabel("Action Performed"), requestResponse);
-
+        /*
+        Request response panel is not editable and only displays the answer.
+         */
+        requestResponse.setEditable(false);
 
         /*
           create new buttons submitDrugRequest and cancelDrugRequest
@@ -84,10 +100,21 @@ public class DrugRequestScreen extends JPanel implements ActionListener, DrugReq
          */
 
         /*
-          create drugRequestButtons panel to add buttons to
+        The drug list added to the JCombo Box to see which are all the types of drugs you can request.
+         */
+        drugs = new  DrugListGenerator().GenerateDrugList();
+        selectDrug = new JComboBox<>(drugs);
+
+
+        /*
+          create drugRequestButtons panel to add buttons to and create teh top panel
          */
         JPanel drugRequestPanel = new JPanel();
+        this.drugRequestPanel = new JPanel();
+        this.drugRequestPanel.setLayout(new FlowLayout());
 
+        this.drugRequestPanel.add(drugNamePanel);
+        this.drugRequestPanel.add(selectDrug);
 //        public JPanel buttonGenerator()
 
         drugRequestPanel.add(drugRequestButton);
@@ -108,7 +135,7 @@ public class DrugRequestScreen extends JPanel implements ActionListener, DrugReq
 
 
         this.add(title);
-        this.add(drugNamePanel);
+        this.add(this.drugRequestPanel);
         this.add(drugBottlePanel);
         this.add(drugRequestPanel);
 //        this.add(setUpButtonListeners(drugRequestButtons));
