+++ src/main/java/user_registration/frameworks_and_drivers/database_generator/UserRegistrationRecorder.java
@@ -1,7 +1,7 @@
 package user_registration.frameworks_and_drivers.database_generator;
 
-import helper_methods.HelperMethods;
-import site_drug_request.drug_request_frameworks_drivers.DrugRequestRecorder;
+import helper_methods.RetrieveLastID;
+import helper_methods.TableHeader;
 import user_registration.use_case.database_access.UserRegistrationDsGateway;
 import user_registration.use_case.database_access.UserRegistrationDsInputData;
 
@@ -97,7 +97,7 @@ public class UserRegistrationRecorder implements UserRegistrationDsGateway {
             userRegistrationWriter = new BufferedWriter(new FileWriter(csvFile, true));
             UserRegistrationDsInputData latestEntry = userRegistrationRequest.get(username);
 
-            int lastRegistrationRequestID = new HelperMethods().LastIDRetriever(2000, csvFile);
+            int lastRegistrationRequestID = new RetrieveLastID().LastIDRetriever(2000, csvFile);
             int siteOrDepotID = UserlocationDetailsReader(latestEntry.getRole());
 
             System.out.println(latestEntry.getRole());
@@ -173,10 +173,10 @@ public class UserRegistrationRecorder implements UserRegistrationDsGateway {
     }
 
     /**
-     * Generate the header of the CSV File
+     * Use the TableHeaderGenerator to make the DrugRequest tables header.
      */
     private void generateHeader(){
-        DrugRequestRecorder.TableGeneratorHelper(csvFile, headers);
+        TableHeader.TableGeneratorHelper(csvFile, headers);
     }
 
     /**
@@ -193,8 +193,6 @@ public class UserRegistrationRecorder implements UserRegistrationDsGateway {
 
             while  ((currentline = bufferedReader.readLine())!= null){
                 data = currentline.split(delimiter);
-                System.out.println(data[3]);
-                System.out.println(reference);
                 if (Objects.equals(data[3].trim(), reference)){
                     return true;
                 }
